Node overhaul - added arbitrary content support
Node now contains two components - a NodeBase and a MetaContent. NodeBase handles drawing the underlying circle of every node. MetaContent is
where tha magic, or rather the mess, is. It is to be generated by a python script which I will write later. It unions together all the models
and actions from various Content components and provides helper update and view functions ridden with case caluses. This is as close as you
can get to polymorphism in Elm, for now. Hopefully in the future suport for type classes will allow for this to be more elegantly rewritten.